fix: bluetape4k-core 코드 리뷰 이슈 수정 및 버그 수정
## bluetape4k-core 주요 수정
- AbstractValueObject: equals에 javaClass 타입 체크 추가
- AtomicIntRoundrobin: ReentrantLock → AtomicInteger CAS 기반으로 교체 (lock-free)
- ExecutorSupport: finally→try-catch 패턴으로 RejectedExecutionException 방지, 오타 수정
- FutureSupport: tryToComplete()의 unreachable isCancelled 분기 제거
- VirtualThreadDispatcher: val 프로퍼티 → fun 함수로 변경 (리소스 누수 방지)
- VirtualThreadReactorScheduler: 동일하게 fun 함수로 변경
- VarargSupport: ClassCastException 방지를 위해 reflect.Array.newInstance 사용
- NumberSupport: StandardNumberTypes를 불변 Set으로 변경
- TimeSpec: NanosPerMillis 상수 오류 수정 (1_000_000L), MILLIS_IN_* Deprecated 처리
- NumberExtensions: millisToNanos() 반환 타입 Int→Long (overflow 방지)
- DurationSupport: year/month 포함 ISO duration 파싱 시 명시적 예외 추가
- ThrowableSupport: getRootCause 순환 참조 무한루프 방지 (visited Set 추가)
- QueueSupport: linkedBlokcingDequeOf/QueueOf 오타 함수 Deprecated 처리
- ProgressionSupport: count() O(n)→O(1) 산술식으로 교체
- KDoc 한국어 보완 및 Apache 유틸리티 함수 다수

## 기타 모듈 수정
- CalendarPeriodCollector: collectingHours에서 분 여부에 따라
  HourRangeCollection/CalendarTimeRange 분기 생성 (테스트 실패 수정)
- FairLockExamples: CompletableDeferred + delay(30ms) 기반 순차 큐잉으로
  Redis FairLock 순서 보장 안정화 (간헐적 실패 수정)
- JdbcRedissonRepository: KDoc 보완

+++ bluetape4k/core/src/main/kotlin/io/bluetape4k/javatimes/NumberExtensions.kt
@@ -396,7 +396,7 @@ fun Long.yearPeriod(): Period = Period.ofYears(this.toInt())
 /**
  * 밀리초를 나노초로 변환합니다.
  */
-fun Long.millisToNanos(): Int = (this * NanosPerMillis).toInt()
+fun Long.millisToNanos(): Long = this * NanosPerMillis
 
 /**
  * 해당 연도가 윤년인지 여부를 반환합니다.
